Remove async-std dependency
Trying to reduce dependencies to the most minimal and standardized set.
Preference given to using the `futures` crate where possible.

For the most part, replacing `async-std` with `futures` was as simple.

The one dependency that couldn't be avoided is some kind of timer for
timeouts.

I ended up keeping `futures-timer`, a general timers crate from the
async-rs project, and then porting the `timeout` module directly from
`async-std`.

I think that whether performance is impacted depends mostly on whether
the `futures_util` equivalents for `async-std` have any shortcomings.
This includes:
- io::copy
- AsyncWriteExt
- AsyncReadExt
- AsyncBufReadExt
- BufReader